fix handling of mutable default params
Remove dictionaries used as default paramaters

I was seeing a situation where the header
'content-type': 'application/x-www-form-urlencoded'
was being set for _all_ requests - not just DELETE

I realised that this was because of the use of dictionaries for default
method parmater values.

http://effbot.org/zone/default-values.htm

I have changed the methods to use a placeholder None value instead.